+++ blog/admin.py
@@ -3,4 +3,4 @@ from . import models
 
 # Register your models here.
 
-admin.site.register(models.Publicacion)
+admin.site.register(models.Post)